browser now: Replace Date.now() with lightstep's implementation
This commit replaces our forked dd-trace-js's browser performance
calculation from using Date.now() to using lightstep's implementation.

One key thing to note is that this will report now in milliseconds
instead of microseconds. This is consistent with our how we report our
metrics (in ms) and should fix the duration reporting inconsistencies
that we have been experiencing in recently. Specifically, we have gotten
reports of negative durations and traces that seem to start in the
future.

The main reason we suspect this "Only use Date.now()" change is that
Date.now() is used for get time relative to Unix epoch time whereas
performance.now() gets the time relative to a page load. We should use
the same method of calculating browser performance across our tools so
that we are consistent.

Going forward, we should standardize our tracing tools to using
performance.now() and defaulting to Date.now() if performance is not
available.